requirements.txt
openai
whisper
gTTS
python-dotenv
